+++ src/str.c
@@ -71,6 +71,7 @@ char **str_split(char *a_str, const char a_delim)
 	return result;
 }
 
+// You must free the result if result is non-NULL.
 const char *str_repeat(char *str, size_t times)
 {
 	if (times < 1)
@@ -83,4 +84,92 @@ const char *str_repeat(char *str, size_t times)
 		strcat(ret, str);
 	}
 	return ret;
+}
+
+// You must free the result if result is non-NULL.
+char *str_replace(char *orig, char *rep, char *with)
+{
+	char *result;
+	char *ins;
+	char *tmp;
+	size_t len_rep;
+	size_t len_with;
+	size_t len_front;
+	size_t count;
+
+	// sanity checks and initialization
+	if (!orig || !rep)
+		return NULL;
+	len_rep = strlen(rep);
+	if (len_rep == 0)
+		return NULL; // empty rep causes infinite loop during count
+	if (!with)
+		with = "";
+	len_with = strlen(with);
+
+	// count the number of replacements needed
+	ins = orig;
+	for (count = 0; (tmp = strstr(ins, rep)); ++count) {
+		ins = tmp + len_rep;
+	}
+
+	tmp = result = malloc(strlen(orig) + (len_with - len_rep) * count + 1);
+
+	if (!result)
+		return NULL;
+
+	// first time through the loop, all the variable are set correctly
+	// from here on,
+	//    tmp points to the end of the result string
+	//    ins points to the next occurrence of rep in orig
+	//    orig points to the remainder of orig after "end of rep"
+	while (count--) {
+		ins = strstr(orig, rep);
+		len_front = (size_t)(ins - orig);
+		tmp = strncpy(tmp, orig, len_front) + len_front;
+		tmp = strcpy(tmp, with) + len_with;
+		orig += len_front + len_rep; // move to next "end of rep"
+	}
+	strcpy(tmp, orig);
+	return result;
+}
+
+char *btkg_str_replace_placeholder(char *input, const char *search,
+				   const char *replace)
+{
+	char *tmp = NULL;
+	tmp = str_replace(input, (char *)search, (char *)replace);
+	if (tmp) {
+		if (input)
+			free(input);
+		return tmp;
+	}
+	return input;
+}
+
+void btkg_str_replace_escape_sequences(char *str)
+{
+	char *read = str;
+	char *write = str;
+	while (*read) {
+		if (*read == '\\' && *(read + 1)) {
+			read++;
+			switch (*read) {
+				case 'n':
+					*write++ = '\n';
+					break;
+				case 't':
+					*write++ = '\t';
+					break;
+				// --
+				default:
+					*write++ = '\\';
+					*write++ = *read;
+			}
+		} else {
+			*write++ = *read;
+		}
+		read++;
+	}
+	*write = '\0';
 }