+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Hostname can be used as a username or password dynamically ($TARGET) (#28)
+- Output log format are now customizable (#30)
 
 ### Fixed
 - Added missing wrapper for FD_SET in static build